id_ed25519

+++ install.sh
@@ -14,9 +14,9 @@ which pacman 2> /dev/null && sudo pacman -Syu --noconfirm git gvim fish openssh
 which apt-get 2> /dev/null && yes|sudo apt-get install git fish vim-gnome
 which yum 2> /dev/null && yes|sudo yum install git fish vim
 
-RM_ID_RSA=false
+RM_ID=false
 
-if [[ ! -d ~/.ssh || ! -f ~/.ssh/id_rsa ]]
+if [[ ! -d ~/.ssh || (! -f ~/.ssh/id_rsa  && ! -f ~.ssh/id_ed25519) ]]
 then
     mkdir -p ~/.ssh
     scp saurelg@ssh.inpt.fr:.ssh/id_rsa .ssh
@@ -24,7 +24,7 @@ then
     ssh-agent -s > ~/.ssh/tmpagent
     . ~/.ssh/tmpagent
     ssh-add
-    RM_ID_RSA=true
+    RM_ID=true
 fi
 
 echo 'ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQDO1DwbWEyl9W9+VxqaIUH4XPVoKMpoxcyh2X9/1NcpTtMmkEwxAfQDONp0R8n4HGc7YpUUFQ0PgjIqhXaOz5zvWwr2wIf+1tPdV3lpxxNawZ8iAwwApaPqLdR0o0NAM8hKTbxB3ObVuzN5T5VJO/r4j1G4kH0wEiOrPnph5LPvkwGlMyIV8B7948v/CAe/YsTQA7jq6oijOAU/MTpjWjXANcDj688IgrDobx9L0T1oAhVrs11SqofrWuWTbgofLIR4mQhbm7t2DXha4kzD82lB9ia6TAXG9mysGsBYkIl2RZ9BA8Ax5ftou1zbtpYyO1SN5hytBi07BsYep/tHCSKn nim@Nausicaa' >> .ssh/authorized_keys
@@ -41,7 +41,7 @@ do
     grep -q $repo .gitrepos 2> /dev/null || echo $HOME/$repo >> .gitrepos
 done
 
-$RM_ID_RSA && rm .ssh/{id_rsa,tmpagent}
+$RM_ID && rm .ssh/{id_rsa,tmpagent}
 
 for file in .bash_profile .bash_logout .tmux.conf .nanorc .xbindkeyrc .vimpagerrc .vimrc .Xdefaults .gitconfig .bashrc .hgrc .zshrc .xmonad .vim .i3 .xinitrc .isort.cfg
 do
